+++ Changelog.md
@@ -2,9 +2,10 @@
 
 ## 3.0.1 - 2026-03-13
 
+- Fix crash on quick note with empty title (Updated to `LinqToOneNote-2.1.1`)
 - Fix incorrect **scope search** check
 - Cache OneNote hierarchy when applicable
-  - This brings the performance improvement for Notebook Explorer from last patch to **recent pages** and **title search**
+  - This brings the performance improvement for Notebook Explorer from last version to **recent pages** and **title search**
 
 ## 3.0.0 - 2026-03-04